Change ingredient amount error handling to float not integer

+++ app/models/beer.rb
@@ -46,7 +46,7 @@ class Beer < ApplicationRecord
     self.beer_ingredients.each do |beer_ingredient|
       if beer_ingredient.amount == nil
         errors.add(:base, "Amount for #{kind_for_amount(beer_ingredient)} can't be blank")
-      elsif beer_ingredient.amount.to_i <= 0
+      elsif beer_ingredient.amount.to_f <= 0
         errors.add(:base, "Amount for #{kind_for_amount(beer_ingredient)} must be a number greater than zero")
       end
     end
